Corrected Generation bug with other games
- Created an Entrances file to put Entrances functions there
- Put all of generate_early(init.py) into set_rules(Rules.py)

Tested a Generation.
Connected to the seed and settings seems right,locations sent correctly
