+++ worlds/apeescape/__init__.py
@@ -7,6 +7,8 @@ from BaseClasses import ItemClassification, MultiWorld, Tutorial, CollectionStat
 from logging import warning
 from Options import OptionError
 from worlds.AutoWorld import WebWorld, World
+
+from .Entrances import get_required_keys, initialize_level_list
 from .Items import item_table, ApeEscapeItem, GROUPED_ITEMS
 from .Locations import location_table, base_location_id, GROUPED_LOCATIONS
 from .Regions import create_regions, ApeEscapeLevel
@@ -56,6 +58,8 @@ class ApeEscapeWorld(World):
 
     item_name_to_id = item_table
 
+    set_rules = set_rules
+
     for key, value in item_name_to_id.items():
         item_name_to_id[key] = value + base_location_id
 
@@ -68,7 +72,7 @@ class ApeEscapeWorld(World):
     location_name_groups = GROUPED_LOCATIONS
 
     def __init__(self, multiworld: MultiWorld, player: int):
-        super().__init__(multiworld, player)
+
         self.goal: Optional[int] = 0
         self.logic: Optional[int] = 0
         self.entrance: Optional[int] = 0
@@ -78,10 +82,10 @@ class ApeEscapeWorld(World):
         self.superflyer: Optional[int] = 0
         self.shufflenet: Optional[int] = 0
         self.shufflewaternet: Optional[int] = 0
-
         self.itempool: List[ApeEscapeItem] = []
         self.levellist: List[ApeEscapeLevel] = []
         self.entranceorder: List[ApeEscapeLevel] = []
+        super(ApeEscapeWorld, self).__init__(multiworld, player)
 
     def generate_early(self) -> None:
         self.goal = self.options.goal.value
@@ -97,24 +101,6 @@ class ApeEscapeWorld(World):
 
     def create_regions(self):
         create_regions(self)
-
-    def set_rules(self):
-        self.levellist = initialize_level_list()
-        # If entrances aren't shuffled, then we don't need to shuffle the entrances.
-        if (self.options.entrance != 0x00):
-            self.random.shuffle(self.levellist)
-            # Some levels need to be kept at a specific entrance - put those back.
-            self.levellist = fixed_levels(self.levellist, self.options.entrance)
-        self.levellist = set_calculated_level_data(self.levellist, self.options.unlocksperkey)
-        # Make a copy of the list for passing to the client for entrance shuffle purposes. We know this list has the levels sorted in the order they'd be presented in-game (so whatever is at the Fossil Field entrance first, etc.)
-        self.entranceorder = list(self.levellist)
-        # If entrances weren't shuffled, then this list is already sorted. We sort the list for ease of setting up access rules in the logic files.
-        if (self.options.entrance != 0x00):
-            self.levellist.sort()
-        set_rules(self)
-
-#    def generate_basic(self) -> None:
-#        generate_basic(self)
 
     def create_item(self, name: str) -> ApeEscapeItem:
         item_id = item_table[name]
@@ -289,61 +275,3 @@ class ApeEscapeWorld(World):
         # filename = f"{self.multiworld.get_out_file_name_base(self.player)}.apae"
         # with open(os.path.join(output_directory, filename), 'w') as f:
         #     json.dump(data, f)
-
-
-def initialize_level_list():
-    levelnames = ["Fossil Field", "Primordial Ooze", "Molten Lava", "Thick Jungle", "Dark Ruins", "Cryptic Relics", "Stadium Attack", "Crabby Beach", "Coral Cave", "Dexter's Island", "Snowy Mammoth", "Frosty Retreat", "Hot Springs", "Gladiator Attack", "Sushi Temple", "Wabi Sabi Wall", "Crumbling Castle", "City Park", "Specter's Factory", "TV Tower", "Monkey Madness", "Peak Point Matrix"]
-    levelids = [0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F, 0x10, 0x11, 0x14, 0x15, 0x16, 0x18, 0x1E]
-    levellist = []
-    for x in range (0, 22):
-        levellist.append(ApeEscapeLevel(levelnames[x], levelids[x], x))
-    return levellist
-
-def level_to_bytes(name):
-    bytelist = []
-    for x in name:
-        bytelist.append(character_lookup(x))
-    return bytelist
-    
-def character_lookup(byte):
-    if byte.isspace(): # Space
-        return 255
-    if byte.isalpha():
-        return ord(byte) - 49 # Both uppercase and lowercase letters
-    if byte.isdecimal():
-        if int(byte) < 6:
-            return ord(byte) + 56 # 0-5
-        else:
-            return ord(byte) + 68 # 6-9
-    if ord(byte) == 39: # Single apostrophe
-        return 187
-
-def fixed_levels(levellist, entoption):
-    for x in range (0, 22):
-        if levellist[x].entrance == 0x1E: # Always reset position of Peak Point Matrix
-            levellist[x], levellist[21] = levellist[21], levellist[x]
-        if levellist[x].entrance == 0x18 and (entoption == 0x01 or entoption == 0x02): # Monkey Madness
-            levellist[x], levellist[20] = levellist[20], levellist[x]
-        if levellist[x].entrance == 0x07 and (entoption == 0x01 or entoption == 0x03): # Stadium Attack
-            levellist[x], levellist[6] = levellist[6], levellist[x]
-        if levellist[x].entrance == 0x0E and (entoption == 0x01 or entoption == 0x03): # Gladiator Attack
-            levellist[x], levellist[13] = levellist[13], levellist[x]
-    return levellist
-
-def set_calculated_level_data(levellist, keyoption):
-    reqkeys = get_required_keys(keyoption)
-    for x in range (0, 22):
-        levellist[x].bytes = level_to_bytes(levellist[x].name)
-        levellist[x].keys = reqkeys[x]
-        levellist[x].newpos = x
-    return levellist
-
-def get_required_keys(option):
-    if option == 0x00: # world
-        return [0,  0,  0,  1,  1,  1,  2,  2,  2,  2,  3,  3,  3,  4,  4,  4,  4,  5,  5,  5,  6,  6]
-    if option == 0x01: # world and races
-        return [0,  0,  0,  1,  1,  1,  2,  3,  3,  3,  4,  4,  4,  5,  6,  6,  6,  7,  7,  7,  8,  8]
-    if option == 0x02: # level
-        return [0,  0,  0,  1,  2,  3,  4,  4,  5,  6,  7,  8,  9,  10, 10, 11, 12, 13, 14, 15, 16, 16]
-    if option == 0x03: # level and races
-        return [0,  0,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9,  10, 11, 12, 13, 14, 15, 16, 17, 18, 18]
